Add Data and Coding Blob variants for data plane
Add `BlobHeader` enum and implement manual serialization/deserialization
from slices

Add `CodingHeader` and `DataHeader` variants

Add methods to convert between `Blob`s and `CodingBlob`s/`DataBlob`s
